+++ src/occsimplifier.h
@@ -305,7 +305,7 @@ private:
     //Persistent data
     Solver*  solver;              ///<The solver this simplifier is connected to
     vector<uint16_t>& seen;
-    vector<uint16_t>& seen2;
+    vector<uint8_t>& seen2;
     vector<Lit>& toClear;
 
     //Temporaries
